test num 22

+++ project_8.py
@@ -84,7 +84,7 @@ class LocationChip:
         data = self.readData()
         print("Type", type(data[1]))
         print("Val", data[1])
-        self.cords = {'a0':int(data[0]), 'a1':int(data[1]), 'a2':int(data[2]), 'a3':int(data[3])} #messing with indexing here
+        self.cords = {'a0':float(data[0]), 'a1':float(data[1]), 'a2':float(data[2]), 'a3':float(data[3])} #messing with indexing here
         self.cords['a3'] = 1000
         closest_cord = min(self.cords, key=self.cords.get)
         print("Current cord:", closest_cord)
